delete old errorPage and add new errorPage
