--- src/components/ErrorPage/ErrorPage.mdx
@@ -1,23 +1,0 @@
-import {Controls, Canvas, Meta, Source} from '@storybook/blocks';
-
-import * as ErrorPageStories from './ErrorPage.stories'
-
-<Meta of={ErrorPageStories} />
-import '../../stories/styles/shared.css';
-
-<div className="header">
-  <h1>ErrorPage</h1>
-  <p>Le composant `ErrorPage` sert à afficher une page d'erreur.</p>
-</div>
-
-<Canvas of={ErrorPageStories.Default} />
-
-# API
-
-<Controls of={ErrorPageStories.Default} />
-
-## Lien de retour
-
-Vous pouvez mettre un lien vers une autre page avec les props `btn-text` et `btn-link` ou `btn-href`.
-
-<Canvas of={ErrorPageStories.WithLink} />
