+++ src/components/ErrorPageNew/index.ts
@@ -1,0 +1,3 @@
+import ErrorPage from './ErrorPage.vue'
+
+export default ErrorPage
