+++ src/components/ErrorPageNew/locales.ts
@@ -1,0 +1,29 @@
+import imgUrlAp from './assets/error-ap.svg'
+
+export const locales = {
+	default: {
+		pageTitle: 'Une erreur est survenue',
+		message: 'Une erreur est survenue de notre côté, veuillez réessayer plus tard.',
+		src: '',
+		supportIdMessage: 'Votre identifiant de support\xa0:',
+		btnText: 'Retour à l’accueil',
+		errorCodeText: 'Code d\'erreur\xa0: ',
+	},
+	cnam: {
+		pageTitle: 'Une erreur est survenue',
+		message: 'Une erreur est survenue de notre côté, veuillez réessayer plus tard. Si le problème persiste veuillez nous contacter par téléphone au 3646',
+		src: '',
+		supportIdMessage: 'Votre identifiant de support\xa0:',
+		btnText: 'Retour à l’accueil',
+		errorCodeText: 'Code d\'erreur\xa0: ',
+	},
+	ap: {
+		pageTitle: 'Les services amelipro sont indisponibles !',
+		message:
+      'Nous mettons en œuvre les mesures nécessaires pour rétablir rapidement votre accès. Veuillez nous excuser pour la gêne occasionnée.',
+		src: imgUrlAp,
+		supportIdMessage: 'Votre identifiant de support\xa0:',
+		btnText: 'Retour à l’accueil',
+		errorCodeText: 'Code d\'erreur\xa0: ',
+	},
+}
